Move middleware exception classes into moin.storage.middleware.exceptions
