+++ src/moin/apps/admin/views.py
@@ -55,11 +55,11 @@ from moin.constants.keys import (
 from moin.constants.namespaces import NAMESPACE_USERPROFILES, NAMESPACE_USERS, NAMESPACE_DEFAULT, NAMESPACE_ALL
 from moin.constants.rights import SUPERUSER, ACL_RIGHTS_CONTENTS, READ, WRITE, CREATE, ADMIN, DESTROY
 from moin.security import require_permission, ACLStringIterator
-from moin.storage.middleware.protecting import AccessDenied
 from moin.utils.crypto import make_uuid
 from moin.datastructures.backends.wiki_groups import WikiGroup
 from moin.datastructures.backends import GroupDoesNotExistError
 from moin.items import Item, acl_validate
+from moin.storage.middleware.exceptions import AccessDenied
 from moin.utils.names import CompositeName, gen_fqnames, parent_names, split_fqname
 from moin.config import default as defaultconfig
 from moin.forms import RequiredText, YourEmail
